Completely Reorganize and Refactor to Edition 2018
upgrades to Rust edition 2018,
moves `digit::BITS` and friends into `Digit::BITS`,
groups contructors and casting related `impl ApInt` functions to a `construction` file
groups data structs and very important related code in the `data` file
groups support structs and related in the `info` file
groups all mathematical and logical `impl ApInt` in `logic` file
removes `Bit` and `.from_bit` in favor of `bool` and `.from_bool`
adds new shift operations and better testing and docs for them,
corrects docs and the ReadMe in many places,
changes the signature of `verify_shift_amount` and `verify_bit_access`
