+++ Cargo.toml
@@ -4,6 +4,7 @@ version = "0.2.0" # Don't forget to update html_root_url in lib.rs.
 authors = ["Robin Freyler <robbepop@web.de>", "Aaron Kutch <aaronkutch@att.net>"]
 license = "MIT/Apache-2.0"
 readme  = "README.md"
+edition = "2018"
 
 keywords   = ["int", "integer", "precision", "arbitrary"]
