+++ README.md
@@ -1,5 +1,4 @@
-ApInt - Arbitrary Precision Integer
-===================================
+# ApInt - Arbitrary Precision Integer
 
 |        Linux        |       Windows       |       Codecov        |       Coveralls      |       Docs       |       Crates.io      |
 |:-------------------:|:-------------------:|:--------------------:|:--------------------:|:----------------:|:--------------------:|
@@ -7,8 +6,7 @@ ApInt - Arbitrary Precision Integer
 
 **Development in progress:** *The implementation has not been finished and may not work.*
 
-**A**rbitrary **p**recision **Int**egers (**ApInt**) represent integers that have an arbitrary but 
-fixed runtime bit-width and offers two's complement modulo arithmetic equal to machine integers.
+**A**rbitrary **p**recision **Int**egers (**ApInt**) represent integers that have an arbitrary but fixed runtime bit-width and offers two's complement modulo arithmetic equal to machine integers.
 
 The integer types offered by this library are:
 
@@ -24,15 +22,22 @@ The API is based on the LLVM [`APInt`](http://llvm.org/doxygen/classllvm_1_1APIn
 - SMT solvers may use this as an underlying model for the theory of bitvectors.
 - Operations and backend for cryptographic keys.
 - Also usable as a simple bitset with runtime length information.
+- Use it like any big integer library, except that the user manages resizes and allocations.
 
 ## Internals
 
 The design focus is at efficiency and robustness.
-`ApInt` instances are small-value-optimized. This means that only `ApInt` instances with a bit-width larger than 64 bits allocate dynamic memory.
+An `ApInt` constists of a sequence of `Digit`s.
+A `Digit` is `u64` by default, but this can be changed TODO
+`ApInt` instances are small-value-optimized. This means that only `ApInt` instances with a bit-width larger than the number of bits in a `Digit` allocate dynamic memory.
 
-An `ApInt` constists of a sequence of 64-bit `Digit`s.
-Computations are done within their 128-bit `DoubleDigit` form to prevent bit-loss on over- or underflows.
-This implies a dependency on 128-bit integers which are currently unstable in Rust.
+Very little `unsafe` is used outside of managing internal `union`s, by default. The robustness of `ApInt` operations is backed by extensive fuzz testing (including unit, regression, random input, and edge case testing in multiple flag modes).
+Internal bounds checking is on by default but can be turned off and `unsafe` indexing turned on by the TODO flag. To be clear, input checks to functions will still be on (e.g. functions will still return `Err`s if input bit widths are not matching). This has the advantage of increasing performance by approximately TODO, but be warned that if a bug is encountered it will cause undefined behavior instead of panicking like it does when the flag is off.
+Even with the TODO flag enabled, `apint` is still arguably one of the safest big integer libraries because of intensive testing with the flag enabled and disabled.
+
+## Performance
+
+For best performance, compile with `O3`, `panic=abort`, `cpu=native` (which can have a decent impact on some CPUs due to better CPU-specific shifting and `.leading_zeros()` functions), TODO (see "Internals" above).
 
 ## Differences & Parallels
 
@@ -43,14 +48,14 @@ The below table lists public and internal differences between `ApInt` and `num::
 | Abstraction              | High-level unbounded integers.            | Twos-complement machine integers.       |
 | Behaviour                | Behaves like an immutable type most often. This results in lots of copies and better usability. | API design with a focus on efficient operations and machine emulation. |
 | Small Value Optimization | No                                        | Yes: Up to 64-bits.                     |
-| Building Blocks          | 32-bit `BigDigit` aka `u32`               | 64-bit `Digit`                          |
+| Building Blocks          | 32-bit `BigDigit` aka `u32`               | 64-bit `Digit` (by default)             |
 | Compute Unit             | 64-bit `DoubleBigDigit` aka `u64`         | 128-bit `DoubleDigit`                   |
 | Signed                   | Yes: `num::BigUint` is for unsigned.      | No: Operations know signedness instead. |
-| `mem::size_of<..>`       | About 24 bytes + some signedness info.    | Exactly 128 bits (16 bytes).            |
+| `mem::size_of<..>`       | About 24 bytes + some signedness info.    | Exactly 128 bits (16 bytes). (this is assuming pointer widths are 64 bits)           |
 | Width interoperability   | No restriction to operate between `BigInt` instances with different bit-widths. | Only `ApInt` instances with the same bit-width can interoperate. |
 | Memory footprint         | Determined by current value stored.       | Determined by bit-width.                |
-| Can grow and shrink?     | Yes                                       | No, see above.                          |
-| Unstable features?       | None                                      | Stable as of Rust 1.26.                 |
+| Grows and shrinks automatically? | Yes                               | No                                      |
+| Unstable features?       | None                                      | Stable as of Rust 1.31.                 |
 
 ## Current State
 
@@ -75,16 +80,16 @@ State of the API modules implemented so far:
 ## Planned Features
 
 - Full and efficient `ApInt` implementation and decent test coverage.
-- Mid-level `ApsInt` wrapper around `ApInt` that stores a run-time sign information.
-  This is different from `Int` and `UInt` since those types store
-  their sign immutable in their type. This is the same as LLVM's `APSInt` data type.
+- Mid-level `ApsInt` wrapper around `ApInt` that stores a run-time sign information. This is different from `Int` and `UInt` since those types store their sign immutable in their type. This is the same as LLVM's `APSInt` data type. These also allow for more efficient multiplication and division operations on negative numbers (see the documentation for those).
+- Low level unsafe functions that have no bounds checking, allow for `ApInt`s of different bit widths to be operated on, and have access to reusing internal allocations for calculations that require allocated temporaries.
+- More efficient operations.
 
 ## License
 
 Licensed under either of
 
- * Apache License, Version 2.0, ([LICENSE-APACHE](LICENSE-APACHE) or http://www.apache.org/licenses/LICENSE-2.0)
- * MIT license ([LICENSE-MIT](LICENSE-MIT) or http://opensource.org/licenses/MIT)
+- Apache License, Version 2.0, ([LICENSE-APACHE](LICENSE-APACHE) or <http://www.apache.org/licenses/LICENSE-2.0>)
+- MIT license ([LICENSE-MIT](LICENSE-MIT) or <http://opensource.org/licenses/MIT>)
 
 at your option.
 
@@ -123,8 +128,19 @@ additional terms or conditions.
 
 ## Release Notes
 
+### Version 0.3.0 TODO
+
+- Removed `Bit`, changed `ApInt::from_bit` to `ApInt::from_bool`
+- Add circular shift functions like `rotate_left_assign`.
+- reorganized internals with updated dependencies followed by `rustfmt`ing and clippy
+- Add the TODO flag.
+- `Digit` can now easily be changed to be a `u32`.
+- Corrected README.md for markdown lints.
+
 ### Version 0.2.0 - 2018-05-16
 
+- Rename many functions from `_checked_` to `_wrapping_` and clarified documentation
+- Added division functions
 - Add `Binary`, `LowerHex` and `UpperHex` impls to `Int`, `UInt` and `ApInt`.  
   Note that implementations for `Octal` are still missing.
