+++ src/data/int.rs
@@ -1,12 +1,6 @@
-use apint::ApInt;
-use traits::Width;
-use digit::Bit;
-use bitwidth::BitWidth;
-use errors::Result;
-use apint::{ShiftAmount};
-use bitpos::{BitPos};
-use int::Int;
-use utils::{try_forward_bin_mut_impl, forward_mut_impl, forward_bin_mut_impl};
+use crate::data::{ApInt, UInt};
+use crate::info::{BitWidth, Result, ShiftAmount, BitPos, Width};
+use crate::logic::{try_forward_bin_mut_impl, forward_mut_impl, forward_bin_mut_impl};
 
 #[cfg(feature = "rand_support")]
 use rand;
@@ -20,6 +14,7 @@ use std::ops::{
     BitAndAssign,
     BitOrAssign,
     BitXorAssign,
+    Neg,
     Add,
     Sub,
     Mul,
@@ -29,227 +24,267 @@ use std::ops::{
     SubAssign,
     MulAssign,
     DivAssign,
-    RemAssign
+    RemAssign,
+    Shl,
+    ShlAssign,
+    Shr,
+    ShrAssign
 };
 
-/// Unsigned machine integer with arbitrary bitwidths and modulo arithmetics.
+/// Signed machine integer with arbitrary bitwidths and modulo arithmetics.
 /// 
-/// Thin convenience wrapper around `ApInt` for static unsigned interpretation of the value.
+/// Thin convenience wrapper around `ApInt` for static signed interpretation of the value.
 /// 
-/// This very cheaply transformes to and from `ApInt` and `Int` instances and together with
-/// `Int` offers a more elegant and higher-level abstraction interface to the lower-level `ApInt`.
+/// This very cheaply transformes to and from `ApInt` and `UInt` instances and together with
+/// `UInt` offers a more elegant and higher-level abstraction interface to the lower-level `ApInt`.
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 #[cfg_attr(serde_derive, Serialize)]
 #[cfg_attr(serde_derive, Deserialize)]
-pub struct UInt {
+pub struct Int {
     value: ApInt,
 }
 
-impl From<ApInt> for UInt {
-    fn from(value: ApInt) -> UInt {
-        UInt { value }
+impl From<ApInt> for Int {
+    fn from(value: ApInt) -> Int {
+        Int { value }
     }
 }
 
-impl UInt {
-    /// Transforms this `UInt` into an equivalent `ApInt` instance.
+impl Int {
+    /// Transforms this `Int` into an equivalent `ApInt` instance.
     pub fn into_apint(self) -> ApInt {
         self.value
     }
 
-    /// Transforms this `UInt` into an equivalent `Int` instance.
-    pub fn into_signed(self) -> Int {
-        Int::from(self.value)
+    /// Transforms this `Int` into an equivalent `UInt` instance.
+    pub fn into_unsigned(self) -> UInt {
+        UInt::from(self.value)
     }
 }
 
 /// # Constructors
-impl UInt {
-    /// Creates a new `UInt` from the given `Bit` value with a bit width of `1`.
+impl Int {
+    /// Creates a new `Int` from the given boolean value with a bit width of `1`.
+    pub fn from_bool(bit: bool) -> Int {
+        Int::from(ApInt::from_bool(bit))
+    }
+
+    /// Creates a new `Int` from a given `i8` value with a bit-width of 8.
+    #[inline]
+    pub fn from_i8(val: i8) -> Int {
+        Int::from(ApInt::from_i8(val))
+    }
+
+    /// Creates a new `Int` from a given `i16` value with a bit-width of 16.
+    #[inline]
+    pub fn from_i16(val: i16) -> Int {
+        Int::from(ApInt::from_i16(val))
+    }
+
+    /// Creates a new `Int` from a given `i32` value with a bit-width of 32.
+    #[inline]
+    pub fn from_i32(val: i32) -> Int {
+        Int::from(ApInt::from_i32(val))
+    }
+
+    /// Creates a new `Int` from a given `i64` value with a bit-width of 64.
+    #[inline]
+    pub fn from_i64(val: i64) -> Int {
+        Int::from(ApInt::from_i64(val))
+    }
+
+    /// Creates a new `Int` from a given `i64` value with a bit-width of 64.
+    pub fn from_i128(val: i128) -> Int {
+        Int::from(ApInt::from_i128(val))
+    }
+
+    /// Creates a new `Int` with the given bit width that represents zero.
+    pub fn zero(width: BitWidth) -> Int {
+        Int::from(ApInt::zero(width))
+    }
+
+    /// Creates a new `Int` with the given bit width that represents one.
+    pub fn one(width: BitWidth) -> Int {
+        Int::from(ApInt::one(width))
+    }
+
+    /// Creates a new `Int` with the given bit width that has all bits unset.
     ///
-    /// This function is generic over types that are convertible to `Bit` such as `bool`.
-    pub fn from_bit<B>(bit: B) -> UInt
-    where
-        B: Into<Bit>,
-    {
-        UInt::from(ApInt::from_bit(bit))
+    /// **Note:** This is equal to calling `Int::zero` with the given `width`.
+    pub fn all_unset(width: BitWidth) -> Int {
+        Int::zero(width)
     }
 
-    /// Creates a new `UInt` from a given `u8` value with a bit-width of 8.
-    #[inline]
-    pub fn from_u8(val: u8) -> UInt {
-        UInt::from(ApInt::from_u8(val))
-    }
-
-    /// Creates a new `UInt` from a given `u16` value with a bit-width of 16.
-    #[inline]
-    pub fn from_u16(val: u16) -> UInt {
-        UInt::from(ApInt::from_u16(val))
-    }
-
-    /// Creates a new `UInt` from a given `u32` value with a bit-width of 32.
-    #[inline]
-    pub fn from_u32(val: u32) -> UInt {
-        UInt::from(ApInt::from_u32(val))
-    }
-
-    /// Creates a new `UInt` from a given `u64` value with a bit-width of 64.
-    #[inline]
-    pub fn from_u64(val: u64) -> UInt {
-        UInt::from(ApInt::from_u64(val))
-    }
-
-    /// Creates a new `UInt` from a given `u64` value with a bit-width of 64.
-    pub fn from_u128(val: u128) -> UInt {
-        UInt::from(ApInt::from_u128(val))
-    }
-
-    /// Creates a new `UInt` with the given bit width that represents zero.
-    pub fn zero(width: BitWidth) -> UInt {
-        UInt::from(ApInt::zero(width))
-    }
-
-    /// Creates a new `UInt` with the given bit width that represents one.
-    pub fn one(width: BitWidth) -> UInt {
-        UInt::from(ApInt::one(width))
-    }
-
-    /// Creates a new `UInt` with the given bit width that has all bits unset.
-    ///
-    /// **Note:** This is equal to calling `UInt::zero` with the given `width`.
-    pub fn all_unset(width: BitWidth) -> UInt {
-        UInt::zero(width)
-    }
-
-    /// Creates a new `UInt` with the given bit width that has all bits set.
+    /// Creates a new `Int` with the given bit width that has all bits set.
     ///
     /// # Note
     ///
     /// - This is equal to minus one on any twos-complement machine.
-    pub fn all_set(width: BitWidth) -> UInt {
-        UInt::from(ApInt::all_set(width))
+    pub fn all_set(width: BitWidth) -> Int {
+        Int::from(ApInt::all_set(width))
     }
 
-    /// Returns the smallest `UInt` that can be represented by the given `BitWidth`.
-    pub fn min_value(width: BitWidth) -> UInt {
-        UInt::from(ApInt::unsigned_min_value(width))
+    /// Returns the smallest `Int` that can be represented by the given `BitWidth`.
+    pub fn min_value(width: BitWidth) -> Int {
+        Int::from(ApInt::signed_min_value(width))
     }
 
-    /// Returns the largest `UInt` that can be represented by the given `BitWidth`.
-    pub fn max_value(width: BitWidth) -> UInt {
-        UInt::from(ApInt::unsigned_max_value(width))
-    }
-}
-
-impl<B> From<B> for UInt
-    where B: Into<Bit>
-{
-    #[inline]
-    fn from(bit: B) -> UInt {
-        UInt::from_bit(bit)
+    /// Returns the largest `Int` that can be represented by the given `BitWidth`.
+    pub fn max_value(width: BitWidth) -> Int {
+        Int::from(ApInt::signed_max_value(width))
     }
 }
 
-impl From<u8> for UInt {
-    fn from(val: u8) -> UInt {
-        UInt::from_u8(val)
+impl From<bool> for Int {
+    fn from(bit: bool) -> Int {
+        Int::from_bool(bit)
     }
 }
 
-impl From<u16> for UInt {
-    fn from(val: u16) -> UInt {
-        UInt::from_u16(val)
+impl From<i8> for Int {
+    fn from(val: i8) -> Int {
+        Int::from_i8(val)
     }
 }
 
-impl From<u32> for UInt {
-    fn from(val: u32) -> UInt {
-        UInt::from_u32(val)
+impl From<i16> for Int {
+    fn from(val: i16) -> Int {
+        Int::from_i16(val)
     }
 }
 
-impl From<u64> for UInt {
-    fn from(val: u64) -> UInt {
-        UInt::from_u64(val)
+impl From<i32> for Int {
+    fn from(val: i32) -> Int {
+        Int::from_i32(val)
     }
 }
 
-impl From<u128> for UInt {
-    fn from(val: u128) -> UInt {
-        UInt::from_u128(val)
+impl From<i64> for Int {
+    fn from(val: i64) -> Int {
+        Int::from_i64(val)
     }
 }
 
-macro_rules! impl_from_array_for_uint {
+impl From<i128> for Int {
+    fn from(val: i128) -> Int {
+        Int::from_i128(val)
+    }
+}
+
+macro_rules! impl_from_array_for_Int {
     ($n:expr) => {
-        impl From<[u64; $n]> for UInt {
-            fn from(val: [u64; $n]) -> UInt {
-                UInt::from(ApInt::from(val))
+        impl From<[i64; $n]> for Int {
+            fn from(val: [i64; $n]) -> Int {
+                Int::from(ApInt::from(val))
             }
         }
     }
 }
 
-impl_from_array_for_uint!(2); // 128 bits
-impl_from_array_for_uint!(3); // 192 bits
-impl_from_array_for_uint!(4); // 256 bits
-impl_from_array_for_uint!(5); // 320 bits
-impl_from_array_for_uint!(6); // 384 bits
-impl_from_array_for_uint!(7); // 448 bits
-impl_from_array_for_uint!(8); // 512 bits
-impl_from_array_for_uint!(16); // 1024 bits
-impl_from_array_for_uint!(32); // 2048 bits
+impl_from_array_for_Int!(2); // 128 bits
+impl_from_array_for_Int!(3); // 192 bits
+impl_from_array_for_Int!(4); // 256 bits
+impl_from_array_for_Int!(5); // 320 bits
+impl_from_array_for_Int!(6); // 384 bits
+impl_from_array_for_Int!(7); // 448 bits
+impl_from_array_for_Int!(8); // 512 bits
+impl_from_array_for_Int!(16); // 1024 bits
+impl_from_array_for_Int!(32); // 2048 bits
 
 /// # Utilities
-impl UInt {
-    /// Returns `true` if this `UInt` represents the value zero (`0`).
+impl Int {
+    /// Returns `true` if this `Int` represents the value zero (`0`).
     ///
     /// # Note
     ///
     /// - Zero (`0`) is also called the additive neutral element.
     /// - This operation is more efficient than comparing two instances
-    ///   of `UInt` for the same reason.
+    ///   of `Int` for the same reason.
     pub fn is_zero(&self) -> bool {
         self.value.is_zero()
     }
 
-    /// Returns `true` if this `UInt` represents the value one (`1`).
+    /// Returns `true` if this `Int` represents the value one (`1`).
     ///
     /// # Note
     ///
     /// - One (`1`) is also called the multiplicative neutral element.
     /// - This operation is more efficient than comparing two instances
-    ///   of `UInt` for the same reason.
+    ///   of `Int` for the same reason.
     pub fn is_one(&self) -> bool {
         self.value.is_one()
     }
 
-    /// Returns `true` if this `UInt` represents an even number.
+    /// Returns `true` if this `Int` represents an even number.
     pub fn is_even(&self) -> bool {
         self.value.is_even()
     }
 
-    /// Returns `true` if this `UInt` represents an odd number.
+    /// Returns `true` if this `Int` represents an odd number.
     pub fn is_odd(&self) -> bool {
         self.value.is_odd()
     }
+
+    /// Returns `true` if the value of this `Int` is positive.
+    pub fn is_positive(&self) -> bool {
+        self.sign_bit() == false
+    }
+
+    /// Returns `true` if the value of this `Int` is negative.
+    pub fn is_negative(&self) -> bool {
+        !self.is_positive()
+    }
+
+    /// Returns a number representing sign of this `ApInt`.
+    /// 
+    /// - `0` if the number is zero
+    /// - `1` if the number is positive
+    /// - `-1` if the number is negative
+    pub fn signum(&self) -> i8 {
+        if self.is_zero() {
+            return 0
+        }
+        if self.is_negative() {
+            return -1
+        }
+        1
+    }
+
+    /// Returns an absolute value representation of this `Int`.
+    /// 
+    /// # Note
+    /// 
+    /// - Consumes `self`.
+    /// - Does nothing for positive `Int` instances.
+    pub fn into_abs(self) -> Int {
+        forward_mut_impl(self, Int::wrapping_abs)
+    }
+
+    /// Converts this `Int` into its absolute value representation.
+    /// 
+    /// - Does nothing for positive `Int` instances.
+    pub fn wrapping_abs(&mut self) {
+        if self.is_negative() {
+            self.wrapping_neg()
+        }
+    }
 }
 
-impl UInt {
-
+/// # Comparisons
+impl Int {
     /// Less-than (`lt`) comparison between `self` and `rhs`.
     /// 
     /// # Note
     /// 
     /// - `checked_` for this function means that it checks the bit widths
     /// - Returns `Ok(true)` if `self < rhs`.
-    /// - Interprets both `UInt` instances as **unsigned** values.
+    /// - Interprets both `Int` instances as **unsigned** values.
     /// 
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn checked_lt(&self, rhs: &UInt) -> Result<bool> {
-        self.value.checked_ult(&rhs.value)
+    pub fn checked_lt(&self, rhs: &Int) -> Result<bool> {
+        self.value.checked_slt(&rhs.value)
     }
 
     /// Less-equals (`le`) comparison between `self` and `rhs`.
@@ -258,14 +293,14 @@ impl UInt {
     /// 
     /// - `checked_` for this function means that it checks the bit widths
     /// - Returns `Ok(true)` if `self <= rhs`.
-    /// - Interprets both `UInt` instances as **unsigned** values.
+    /// - Interprets both `Int` instances as **unsigned** values.
     /// 
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
     #[inline]
-    pub fn checked_le(&self, rhs: &UInt) -> Result<bool> {
-        self.value.checked_ule(&rhs.value)
+    pub fn checked_le(&self, rhs: &Int) -> Result<bool> {
+        self.value.checked_sle(&rhs.value)
     }
 
     /// Greater-than (`gt`) comparison between `self` and `rhs`.
@@ -274,14 +309,14 @@ impl UInt {
     /// 
     /// - `checked_` for this function means that it checks the bit widths
     /// - Returns `Ok(true)` if `self > rhs`.
-    /// - Interprets both `UInt` instances as **unsigned** values.
+    /// - Interprets both `Int` instances as **unsigned** values.
     /// 
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
     #[inline]
-    pub fn checked_gt(&self, rhs: &UInt) -> Result<bool> {
-        self.value.checked_ugt(&rhs.value)
+    pub fn checked_gt(&self, rhs: &Int) -> Result<bool> {
+        self.value.checked_sgt(&rhs.value)
     }
 
     /// Greater-equals (`ge`) comparison between `self` and `rhs`.
@@ -290,21 +325,21 @@ impl UInt {
     /// 
     /// - `checked_` for this function means that it checks the bit widths
     /// - Returns `Ok(true)` if `self >= rhs`.
-    /// - Interprets both `UInt` instances as **unsigned** values.
+    /// - Interprets both `Int` instances as **unsigned** values.
     /// 
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
     #[inline]
-    pub fn checked_ge(&self, rhs: &UInt) -> Result<bool> {
-        self.value.checked_uge(&rhs.value)
+    pub fn checked_ge(&self, rhs: &Int) -> Result<bool> {
+        self.value.checked_sge(&rhs.value)
     }
 }
 
 /// If `self` and `rhs` have unmatching bit widths, `None` will be returned for `partial_cmp`
 /// and `false` will be returned for the rest of the `PartialOrd` methods.
-impl PartialOrd for UInt {
-    fn partial_cmp(&self, rhs: &UInt) -> Option<Ordering> {
+impl PartialOrd for Int {
+    fn partial_cmp(&self, rhs: &Int) -> Option<Ordering> {
         if self.value.width() != rhs.value.width() {
             return None;
         }
@@ -317,271 +352,269 @@ impl PartialOrd for UInt {
         Some(Ordering::Greater)
     }
 
-    fn lt(&self, rhs: &UInt) -> bool {
+    fn lt(&self, rhs: &Int) -> bool {
         self.checked_lt(rhs).unwrap_or(false)
     }
 
-    fn le(&self, rhs: &UInt) -> bool {
+    fn le(&self, rhs: &Int) -> bool {
         self.checked_le(rhs).unwrap_or(false)
     }
 
-    fn gt(&self, rhs: &UInt) -> bool {
+    fn gt(&self, rhs: &Int) -> bool {
         self.checked_gt(rhs).unwrap_or(false)
     }
 
-    fn ge(&self, rhs: &UInt) -> bool {
+    fn ge(&self, rhs: &Int) -> bool {
         self.checked_ge(rhs).unwrap_or(false)
     }
 }
 
 /// # To Primitive (Resize)
-impl UInt {
-    /// Resizes this `UInt` to a `bool` primitive type.
+impl Int {
+    /// Resizes this `Int` to a `bool` primitive type.
     ///
-    /// Bits in this `UInt` that are not within the bounds
+    /// Bits in this `Int` that are not within the bounds
     /// of the `bool` are being ignored.
     ///
     /// # Note
     ///
     /// - Basically this returns `true` if the least significant
-    ///   bit of this `UInt` is `1` and `false` otherwise.
+    ///   bit of this `Int` is `1` and `false` otherwise.
     pub fn resize_to_bool(&self) -> bool {
         self.value.resize_to_bool()
     }
 
-    /// Resizes this `UInt` to a `u8` primitive type.
+    /// Resizes this `Int` to a `i8` primitive type.
     ///
     /// # Note
     ///
     /// - All bits but the least significant `8` bits are
     ///   being ignored by this operation to construct the
     ///   result.
-    pub fn resize_to_u8(&self) -> u8 {
-        self.value.resize_to_u8()
+    pub fn resize_to_i8(&self) -> i8 {
+        self.value.resize_to_i8()
     }
 
-    /// Resizes this `UInt` to a `u16` primitive type.
+    /// Resizes this `Int` to a `i16` primitive type.
     ///
     /// # Note
     ///
     /// - All bits but the least significant `16` bits are
     ///   being ignored by this operation to construct the
     ///   result.
-    pub fn resize_to_u16(&self) -> u16 {
-        self.value.resize_to_u16()
+    pub fn resize_to_i16(&self) -> i16 {
+        self.value.resize_to_i16()
     }
 
-    /// Resizes this `UInt` to a `u32` primitive type.
+    /// Resizes this `Int` to a `i32` primitive type.
     ///
     /// # Note
     ///
     /// - All bits but the least significant `32` bits are
     ///   being ignored by this operation to construct the
     ///   result.
-    pub fn resize_to_u32(&self) -> u32 {
-        self.value.resize_to_u32()
+    pub fn resize_to_i32(&self) -> i32 {
+        self.value.resize_to_i32()
     }
 
-    /// Resizes this `UInt` to a `u64` primitive type.
+    /// Resizes this `Int` to a `i64` primitive type.
     ///
     /// # Note
     ///
     /// - All bits but the least significant `64` bits are
     ///   being ignored by this operation to construct the
     ///   result.
-    pub fn resize_to_u64(&self) -> u64 {
-        self.value.resize_to_u64()
+    pub fn resize_to_i64(&self) -> i64 {
+        self.value.resize_to_i64()
     }
 
-    /// Resizes this `UInt` to a `u128` primitive type.
+    /// Resizes this `Int` to a `i128` primitive type.
     ///
     /// # Note
     ///
     /// - All bits but the least significant `128` bits are
     ///   being ignored by this operation to construct the
     ///   result.
-    pub fn resize_to_u128(&self) -> u128 {
-        self.value.resize_to_u128()
+    pub fn resize_to_i128(&self) -> i128 {
+        self.value.resize_to_i128()
     }
 }
 
 /// # To Primitive (Try-Cast)
-impl UInt {
-    /// Tries to represent the value of this `UInt` as a `bool`.
+impl Int {
+    /// Tries to represent the value of this `Int` as a `bool`.
     ///
     /// # Note
     ///
-    /// This returns `true` if the value represented by this `UInt`
+    /// This returns `true` if the value represented by this `Int`
     /// is `1`, returns `false` if the value represented by this
-    /// `UInt` is `0` and returns an error otherwise.
+    /// `Int` is `0` and returns an error otherwise.
     ///
     /// # Errors
     ///
-    /// - If the value represented by this `UInt` can not be
+    /// - If the value represented by this `Int` can not be
     ///   represented by a `bool`.
     pub fn try_to_bool(&self) -> Result<bool> {
         self.value.try_to_bool()
     }
 
-    /// Tries to represent the value of this `UInt` as a `u8`.
+    /// Tries to represent the value of this `Int` as a `i8`.
     ///
     /// # Note
     ///
     /// - This conversion is possible as long as the value represented
-    ///   by this `UInt` does not exceed the maximum value of `u8`.
+    ///   by this `Int` does not exceed the maximum value of `i8`.
     ///
     /// # Errors
     ///
-    /// - If the value represented by this `UInt` can not be
+    /// - If the value represented by this `Int` can not be
     ///   represented by a `u8`.
-    pub fn try_to_u8(&self) -> Result<u8> {
-        self.value.try_to_u8()
+    pub fn try_to_i8(&self) -> Result<i8> {
+        self.value.try_to_i8()
     }
 
-    /// Tries to represent the value of this `UInt` as a `u16`.
+    /// Tries to represent the value of this `Int` as a `i16`.
     ///
     /// # Note
     ///
     /// - This conversion is possible as long as the value represented
-    ///   by this `UInt` does not exceed the maximum value of `u16`.
+    ///   by this `Int` does not exceed the maximum value of `i16`.
     ///
     /// # Errors
     ///
-    /// - If the value represented by this `UInt` can not be
-    ///   represented by a `u16`.
-    pub fn try_to_u16(&self) -> Result<u16> {
-        self.value.try_to_u16()
+    /// - If the value represented by this `Int` can not be
+    ///   represented by a `i16`.
+    pub fn try_to_i16(&self) -> Result<i16> {
+        self.value.try_to_i16()
     }
 
-    /// Tries to represent the value of this `UInt` as a `u32`.
+    /// Tries to represent the value of this `Int` as a `i32`.
     ///
     /// # Note
     ///
     /// - This conversion is possible as long as the value represented
-    ///   by this `UInt` does not exceed the maximum value of `u32`.
+    ///   by this `Int` does not exceed the maximum value of `i32`.
     ///
     /// # Errors
     ///
-    /// - If the value represented by this `UInt` can not be
-    ///   represented by a `u32`.
-    pub fn try_to_u32(&self) -> Result<u32> {
-        self.value.try_to_u32()
+    /// - If the value represented by this `Int` can not be
+    ///   represented by a `i32`.
+    pub fn try_to_i32(&self) -> Result<i32> {
+        self.value.try_to_i32()
     }
 
-    /// Tries to represent the value of this `UInt` as a `u64`.
+    /// Tries to represent the value of this `Int` as a `i64`.
     ///
     /// # Note
     ///
     /// - This conversion is possible as long as the value represented
-    ///   by this `UInt` does not exceed the maximum value of `u64`.
+    ///   by this `Int` does not exceed the maximum value of `i64`.
     ///
     /// # Errors
     ///
-    /// - If the value represented by this `UInt` can not be
-    ///   represented by a `u64`.
-    pub fn try_to_u64(&self) -> Result<u64> {
-        self.value.try_to_u64()
+    /// - If the value represented by this `Int` can not be
+    ///   represented by a `i64`.
+    pub fn try_to_i64(&self) -> Result<i64> {
+        self.value.try_to_i64()
     }
 
-    /// Tries to represent the value of this `UInt` as a `u128`.
+    /// Tries to represent the value of this `Int` as a `i128`.
     ///
     /// # Note
     ///
     /// - This conversion is possible as long as the value represented
-    ///   by this `UInt` does not exceed the maximum value of `u128`.
+    ///   by this `Int` does not exceed the maximum value of `i128`.
     ///
     /// # Complexity
     ///
-    /// - 𝒪(n) where n is the number of digits of this `UInt`.
+    /// - 𝒪(n) where n is the number of digits of this `Int`.
     ///
     /// # Errors
     ///
-    /// - If the value represented by this `UInt` can not be
-    ///   represented by a `u128`.
-    pub fn try_to_u128(&self) -> Result<u128> {
-        self.value.try_to_u128()
+    /// - If the value represented by this `Int` can not be
+    ///   represented by a `i128`.
+    pub fn try_to_i128(&self) -> Result<i128> {
+        self.value.try_to_i128()
     }
 }
 
 /// # Shifts
-impl UInt {
-       /// Shift this `UInt` left by the given `shift_amount` bits.
+impl Int {
+    /// Shift this `Int` left by the given `shift_amount` bits.
     /// 
     /// This operation is inplace and will **not** allocate memory.
     /// 
     /// # Errors
     /// 
-    /// - If the given `shift_amount` is invalid for the bit width of this `UInt`.
+    /// - If the given `shift_amount` is invalid for the bit width of this `Int`.
     pub fn wrapping_shl_assign<S>(&mut self, shift_amount: S) -> Result<()>
         where S: Into<ShiftAmount>
     {
         self.value.wrapping_shl_assign(shift_amount)
     }
 
-    /// Shift this `UInt` left by the given `shift_amount` bits and returns the result.
+    /// Shift this `Int` left by the given `shift_amount` bits and returns the result.
     /// 
     /// This operation is inplace and will **not** allocate memory.
     /// 
     /// # Errors
     /// 
-    /// - If the given `shift_amount` is invalid for the bit width of this `UInt`.
-    pub fn into_wrapping_shl<S>(self, shift_amount: S) -> Result<UInt>
+    /// - If the given `shift_amount` is invalid for the bit width of this `Int`.
+    pub fn into_wrapping_shl<S>(self, shift_amount: S) -> Result<Int>
         where S: Into<ShiftAmount>
     {
-        self.value.into_wrapping_shl(shift_amount).map(UInt::from)
+        self.value.into_wrapping_shl(shift_amount).map(Int::from)
     }
 
-    /// Right-shifts this `UInt` by the given `shift_amount` bits.
+    /// Right-shifts this `Int` by the given `shift_amount` bits.
     /// 
     /// This operation is inplace and will **not** allocate memory.
     /// 
     /// # Errors
     /// 
-    /// - If the given `shift_amount` is invalid for the bit width of this `UInt`.
+    /// - If the given `shift_amount` is invalid for the bit width of this `Int`.
     pub fn wrapping_shr_assign<S>(&mut self, shift_amount: S) -> Result<()>
         where S: Into<ShiftAmount>
     {
-        self.value.wrapping_lshr_assign(shift_amount)
+        self.value.wrapping_ashr_assign(shift_amount)
     }
 
-    /// Right-shifts this `UInt` by the given `shift_amount` bits
+    /// Right-shifts this `Int` by the given `shift_amount` bits
     /// and returns the result.
     /// 
     /// This operation is inplace and will **not** allocate memory.
     /// 
     /// # Errors
     /// 
-    /// - If the given `shift_amount` is invalid for the bit width of this `UInt`.
-    pub fn into_wrapping_shr<S>(self, shift_amount: S) -> Result<UInt>
+    /// - If the given `shift_amount` is invalid for the bit width of this `Int`.
+    pub fn into_wrapping_shr<S>(self, shift_amount: S) -> Result<Int>
         where S: Into<ShiftAmount>
     {
-        self.value.into_wrapping_lshr(shift_amount).map(UInt::from)
+        self.value.into_wrapping_ashr(shift_amount).map(Int::from)
     }
 }
 
-use std::ops::{Shl, ShlAssign, Shr, ShrAssign};
-
-impl<S> Shl<S> for UInt
+impl<S> Shl<S> for Int
     where S: Into<ShiftAmount>
 {
-    type Output = UInt;
+    type Output = Int;
 
     fn shl(self, shift_amount: S) -> Self::Output {
         self.into_wrapping_shl(shift_amount).unwrap()
     }
 }
 
-impl<S> Shr<S> for UInt
+impl<S> Shr<S> for Int
     where S: Into<ShiftAmount>
 {
-    type Output = UInt;
+    type Output = Int;
 
     fn shr(self, shift_amount: S) -> Self::Output {
         self.into_wrapping_shr(shift_amount).unwrap()
     }
 }
 
-impl<S> ShlAssign<S> for UInt
+impl<S> ShlAssign<S> for Int
     where S: Into<ShiftAmount>
 {
     fn shl_assign(&mut self, shift_amount: S) {
@@ -589,7 +622,7 @@ impl<S> ShlAssign<S> for UInt
     }
 }
 
-impl<S> ShrAssign<S> for UInt
+impl<S> ShrAssign<S> for Int
     where S: Into<ShiftAmount>
 {
     fn shr_assign(&mut self, shift_amount: S) {
@@ -599,30 +632,30 @@ impl<S> ShrAssign<S> for UInt
 
 /// # Random Utilities using `rand` crate.
 #[cfg(feature = "rand_support")]
-impl UInt {
-    /// Creates a new `UInt` with the given `BitWidth` and random `Digit`s.
-    pub fn random_with_width(width: BitWidth) -> UInt {
-        UInt::from(ApInt::random_with_width(width))
+impl Int {
+    /// Creates a new `Int` with the given `BitWidth` and random `Digit`s.
+    pub fn random_with_width(width: BitWidth) -> Int {
+        Int::from(ApInt::random_with_width(width))
     }
 
-    /// Creates a new `UInt` with the given `BitWidth` and random `Digit`s
+    /// Creates a new `Int` with the given `BitWidth` and random `Digit`s
     /// using the given random number generator.
     /// 
     /// **Note:** This is useful for cryptographic or testing purposes.
-    pub fn random_with_width_using<R>(width: BitWidth, rng: &mut R) -> UInt
+    pub fn random_with_width_using<R>(width: BitWidth, rng: &mut R) -> Int
         where R: rand::Rng
     {
-        UInt::from(ApInt::random_with_width_using(width, rng))
+        Int::from(ApInt::random_with_width_using(width, rng))
     }
 
-    /// Randomizes the digits of this `UInt` inplace.
+    /// Randomizes the digits of this `Int` inplace.
     /// 
     /// This won't change its `BitWidth`.
     pub fn randomize(&mut self) {
         self.value.randomize()
     }
 
-    /// Randomizes the digits of this `UInt` inplace using the given
+    /// Randomizes the digits of this `Int` inplace using the given
     /// random number generator.
     /// 
     /// This won't change its `BitWidth`.
@@ -633,18 +666,18 @@ impl UInt {
     }
 }
 
-impl UInt {
-    /// Assigns `rhs` to this `UInt`.
+impl Int {
+    /// Assigns `rhs` to this `Int`.
     ///
     /// This mutates digits and may affect the bitwidth of `self`
     /// which **might result in an expensive operations**.
     ///
     /// After this operation `rhs` and `self` are equal to each other.
-    pub fn assign(&mut self, rhs: &UInt) {
+    pub fn assign(&mut self, rhs: &Int) {
         self.value.assign(&rhs.value)
     }
 
-    /// Strictly assigns `rhs` to this `UInt`.
+    /// Strictly assigns `rhs` to this `Int`.
     /// 
     /// After this operation `rhs` and `self` are equal to each other.
     /// 
@@ -655,32 +688,32 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `rhs` and `self` have unmatching bit widths.
-    pub fn strict_assign(&mut self, rhs: &UInt) -> Result<()> {
+    pub fn strict_assign(&mut self, rhs: &Int) -> Result<()> {
         self.value.strict_assign(&rhs.value)
     }
 }
 
 /// # Casting: Truncation & Extension
-impl UInt {
-    /// Tries to truncate this `UInt` inplace to the given `target_width`
+impl Int {
+    /// Tries to truncate this `Int` inplace to the given `target_width`
     /// and returns the result.
     /// 
     /// # Note
     /// 
     /// - This is useful for method chaining.
     /// - For more details look into
-    ///   [`truncate`](struct.UInt.html#method.truncate).
+    ///   [`truncate`](struct.Int.html#method.truncate).
     /// 
     /// # Errors
     /// 
     /// - If the `target_width` is greater than the current width.
-    pub fn into_truncate<W>(self, target_width: W) -> Result<UInt>
+    pub fn into_truncate<W>(self, target_width: W) -> Result<Int>
         where W: Into<BitWidth>
     {
-        try_forward_bin_mut_impl(self, target_width, UInt::truncate)
+        try_forward_bin_mut_impl(self, target_width, Int::truncate)
     }
 
-    /// Tries to truncate this `UInt` inplace to the given `target_width`.
+    /// Tries to truncate this `Int` inplace to the given `target_width`.
     /// 
     /// # Note
     /// 
@@ -699,25 +732,25 @@ impl UInt {
 
     // ========================================================================
 
-    /// Tries to zero-extend this `UInt` inplace to the given `target_width`
+    /// Tries to zero-extend this `Int` inplace to the given `target_width`
     /// and returns the result.
     /// 
     /// # Note
     /// 
     /// - This is useful for method chaining.
     /// - For more details look into
-    ///   [`extend`](struct.UInt.html#method.extend).
+    ///   [`extend`](struct.Int.html#method.extend).
     /// 
     /// # Errors
     /// 
     /// - If the `target_width` is less than the current width.
-    pub fn into_extend<W>(self, target_width: W) -> Result<UInt>
+    pub fn into_extend<W>(self, target_width: W) -> Result<Int>
         where W: Into<BitWidth>
     {
-        try_forward_bin_mut_impl(self, target_width, UInt::extend)
+        try_forward_bin_mut_impl(self, target_width, Int::extend)
     }
 
-    /// Tries to extend this `UInt` inplace to the given `target_width`.
+    /// Tries to extend this `Int` inplace to the given `target_width`.
     /// 
     /// # Note
     /// 
@@ -731,121 +764,121 @@ impl UInt {
     pub fn extend<W>(&mut self, target_width: W) -> Result<()>
         where W: Into<BitWidth>
     {
-        self.value.zero_extend(target_width)
+        self.value.sign_extend(target_width)
     }
 
     // ========================================================================
 
-    /// Resizes this `UInt` to the given `target_width`
+    /// Resizes this `Int` to the given `target_width`
     /// and returns the result.
     /// 
     /// # Note
     /// 
     /// - This is useful for method chaining.
     /// - For more details look into
-    ///   [`resize`](struct.UInt.html#method.resize).
-    pub fn into_resize<W>(self, target_width: W) -> UInt
+    ///   [`resize`](struct.Int.html#method.resize).
+    pub fn into_resize<W>(self, target_width: W) -> Int
         where W: Into<BitWidth>
     {
-        forward_bin_mut_impl(self, target_width, UInt::resize)
+        forward_bin_mut_impl(self, target_width, Int::resize)
     }
 
-    /// Resizes the given `UInt` inplace.
+    /// Resizes the given `Int` inplace.
     /// 
     /// # Note
     /// 
     /// This operation will forward to
     /// 
-    /// - [`truncate`](struct.UInt.html#method.truncate)
+    /// - [`truncate`](struct.Int.html#method.truncate)
     ///   if `target_width` is less than or equal to the width of
-    ///   the given `UInt`
-    /// - [`extend`](struct.UInt.html#method.extend)
+    ///   the given `Int`
+    /// - [`extend`](struct.Int.html#method.extend)
     ///   otherwise
     pub fn resize<W>(&mut self, target_width: W)
         where W: Into<BitWidth>
     {
-        self.value.zero_resize(target_width)
+        self.value.sign_resize(target_width)
     }
 }
 
 /// # Bitwise Operations
-impl UInt {
+impl Int {
     /// Flips all bits of `self` and returns the result.
     pub fn into_bitnot(self) -> Self {
-        forward_mut_impl(self, UInt::bitnot)
+        forward_mut_impl(self, Int::bitnot)
     }
 
-    /// Flip all bits of this `UInt` inplace.
+    /// Flip all bits of this `Int` inplace.
     pub fn bitnot(&mut self) {
         self.value.bitnot()
     }
 
-    /// Tries to bit-and assign this `UInt` inplace to `rhs`
+    /// Tries to bit-and assign this `Int` inplace to `rhs`
     /// and returns the result.
     /// 
     /// # Errors
     /// 
     /// If `self` and `rhs` have unmatching bit widths.
-    pub fn into_bitand(self, rhs: &UInt) -> Result<UInt> {
-        try_forward_bin_mut_impl(self, rhs, UInt::bitand_assign)
+    pub fn into_bitand(self, rhs: &Int) -> Result<Int> {
+        try_forward_bin_mut_impl(self, rhs, Int::bitand_assign)
     }
 
-    /// Bit-and assigns all bits of this `UInt` with the bits of `rhs`.
+    /// Bit-and assigns all bits of this `Int` with the bits of `rhs`.
     /// 
     /// **Note:** This operation is inplace of `self` and won't allocate memory.
     /// 
     /// # Errors
     /// 
     /// If `self` and `rhs` have unmatching bit widths.
-    pub fn bitand_assign(&mut self, rhs: &UInt) -> Result<()> {
+    pub fn bitand_assign(&mut self, rhs: &Int) -> Result<()> {
         self.value.bitand_assign(&rhs.value)
     }
 
-    /// Tries to bit-and assign this `UInt` inplace to `rhs`
+    /// Tries to bit-and assign this `Int` inplace to `rhs`
     /// and returns the result.
     /// 
     /// # Errors
     /// 
     /// If `self` and `rhs` have unmatching bit widths.
-    pub fn into_bitor(self, rhs: &UInt) -> Result<UInt> {
-        try_forward_bin_mut_impl(self, rhs, UInt::bitor_assign)
+    pub fn into_bitor(self, rhs: &Int) -> Result<Int> {
+        try_forward_bin_mut_impl(self, rhs, Int::bitor_assign)
     }
 
-    /// Bit-or assigns all bits of this `UInt` with the bits of `rhs`.
+    /// Bit-or assigns all bits of this `Int` with the bits of `rhs`.
     /// 
     /// **Note:** This operation is inplace of `self` and won't allocate memory.
     /// 
     /// # Errors
     /// 
     /// If `self` and `rhs` have unmatching bit widths.
-    pub fn bitor_assign(&mut self, rhs: &UInt) -> Result<()> {
+    pub fn bitor_assign(&mut self, rhs: &Int) -> Result<()> {
         self.value.bitor_assign(&rhs.value)
     }
 
-    /// Tries to bit-xor assign this `UInt` inplace to `rhs`
+    /// Tries to bit-xor assign this `Int` inplace to `rhs`
     /// and returns the result.
     /// 
     /// # Errors
     /// 
     /// If `self` and `rhs` have unmatching bit widths.
-    pub fn into_bitxor(self, rhs: &UInt) -> Result<UInt> {
-        try_forward_bin_mut_impl(self, rhs, UInt::bitxor_assign)
+    pub fn into_bitxor(self, rhs: &Int) -> Result<Int> {
+        try_forward_bin_mut_impl(self, rhs, Int::bitxor_assign)
     }
 
-    /// Bit-xor assigns all bits of this `UInt` with the bits of `rhs`.
+    /// Bit-xor assigns all bits of this `Int` with the bits of `rhs`.
     /// 
     /// **Note:** This operation is inplace of `self` and won't allocate memory.
     /// 
     /// # Errors
     /// 
     /// If `self` and `rhs` have unmatching bit widths.
-    pub fn bitxor_assign(&mut self, rhs: &UInt) -> Result<()> {
+    pub fn bitxor_assign(&mut self, rhs: &Int) -> Result<()> {
         self.value.bitxor_assign(&rhs.value)
     }
 }
 
 /// # Bitwise Access
-impl UInt {
+impl Int {
     /// Returns the bit at the given bit position `pos`.
     /// 
     /// This returns
@@ -855,8 +888,8 @@ impl UInt {
     /// 
     /// # Errors
     /// 
-    /// - If `pos` is not a valid bit position for the width of this `UInt`.
-    pub fn get_bit_at<P>(&self, pos: P) -> Result<Bit>
+    /// - If `pos` is not a valid bit position for the width of this `Int`.
+    pub fn get_bit_at<P>(&self, pos: P) -> Result<bool>
         where P: Into<BitPos>
     {
         self.value.get_bit_at(pos)
@@ -866,7 +899,7 @@ impl UInt {
     /// 
     /// # Errors
     /// 
-    /// - If `pos` is not a valid bit position for the width of this `UInt`.
+    /// - If `pos` is not a valid bit position for the width of this `Int`.
     pub fn set_bit_at<P>(&mut self, pos: P) -> Result<()>
         where P: Into<BitPos>
     {
@@ -877,7 +910,7 @@ impl UInt {
     /// 
     /// # Errors
     /// 
-    /// - If `pos` is not a valid bit position for the width of this `UInt`.
+    /// - If `pos` is not a valid bit position for the width of this `Int`.
     pub fn unset_bit_at<P>(&mut self, pos: P) -> Result<()>
         where P: Into<BitPos>
     {
@@ -893,57 +926,85 @@ impl UInt {
     /// 
     /// # Errors
     /// 
-    /// - If `pos` is not a valid bit position for the width of this `UInt`.
+    /// - If `pos` is not a valid bit position for the width of this `Int`.
     pub fn flip_bit_at<P>(&mut self, pos: P) -> Result<()>
         where P: Into<BitPos>
     {
         self.value.flip_bit_at(pos)
     }
 
-    /// Sets all bits of this `UInt` to one (`1`).
+    /// Sets all bits of this `Int` to one (`1`).
     pub fn set_all(&mut self) {
         self.value.set_all()
     }
 
-    /// Returns `true` if all bits in this `UInt` are set.
+    /// Returns `true` if all bits in this `Int` are set.
     pub fn is_all_set(&self) -> bool {
         self.value.is_all_set()
     }
 
-    /// Sets all bits of this `UInt` to zero (`0`).
+    /// Sets all bits of this `Int` to zero (`0`).
     pub fn unset_all(&mut self) {
         self.value.unset_all()
     }
 
-    /// Returns `true` if all bits in this `UInt` are unset.
+    /// Returns `true` if all bits in this `Int` are unset.
     pub fn is_all_unset(&self) -> bool {
         self.value.is_all_unset()
     }
 
-    /// Flips all bits of this `UInt`.
+    /// Flips all bits of this `Int`.
     pub fn flip_all(&mut self) {
         self.value.flip_all()
+    }
+
+    /// Returns the sign bit of this `Int`.
+    /// 
+    /// **Note:** This is equal to the most significant bit of this `Int`.
+    pub fn sign_bit(&self) -> bool {
+        self.value.sign_bit()
+    }
+
+    /// Sets the sign bit of this `Int` to one (`1`).
+    pub fn set_sign_bit(&mut self) {
+        self.value.set_sign_bit()
+    }
+
+    /// Sets the sign bit of this `Int` to zero (`0`).
+    pub fn unset_sign_bit(&mut self) {
+        self.value.unset_sign_bit()
+    }
+
+    /// Flips the sign bit of this `Int`.
+    /// 
+    /// # Note
+    /// 
+    /// - If the sign bit was `0` it will be `1` after this operation and vice versa.
+    /// - Depending on the interpretation of the `Int` this
+    ///   operation changes its signedness.
+    pub fn flip_sign_bit(&mut self) {
+        self.value.flip_sign_bit()
     }
 }
 
 /// # Bitwise utility methods.
-impl UInt {
-    /// Returns the number of ones in the binary representation of this `UInt`.
+impl Int {
+    /// Returns the number of ones in the binary representation of this `Int`.
     pub fn count_ones(&self) -> usize {
         self.value.count_ones()
     }
 
-    /// Returns the number of zeros in the binary representation of this `UInt`.
+    /// Returns the number of zeros in the binary representation of this `Int`.
     pub fn count_zeros(&self) -> usize {
         self.value.count_zeros()
     }
 
-    /// Returns the number of leading zeros in the binary representation of this `UInt`.
+    /// Returns the number of leading zeros in the binary representation of this `Int`.
     pub fn leading_zeros(&self) -> usize {
         self.value.leading_zeros()
     }
 
-    /// Returns the number of trailing zeros in the binary representation of this `UInt`.
+    /// Returns the number of trailing zeros in the binary representation of this `Int`.
     pub fn trailing_zeros(&self) -> usize {
         self.value.trailing_zeros()
     }
@@ -953,11 +1014,11 @@ impl UInt {
 //  `Not` (bitwise) impls
 //  ===========================================================================
 
-impl Not for UInt {
-    type Output = UInt;
+impl Not for Int {
+    type Output = Int;
 
     fn not(self) -> Self::Output {
-        forward_mut_impl(self, UInt::bitnot)
+        forward_mut_impl(self, Int::bitnot)
     }
 }
 
@@ -965,26 +1026,26 @@ impl Not for UInt {
 //  `BitAnd` impls
 //  ===========================================================================
 
-impl<'a> BitAnd<&'a UInt> for UInt {
-    type Output = UInt;
+impl<'a> BitAnd<&'a Int> for Int {
+    type Output = Int;
 
-    fn bitand(self, rhs: &'a UInt) -> Self::Output {
+    fn bitand(self, rhs: &'a Int) -> Self::Output {
         self.into_bitand(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> BitAnd<&'a UInt> for &'b UInt {
-    type Output = UInt;
+impl<'a, 'b> BitAnd<&'a Int> for &'b Int {
+    type Output = Int;
 
-    fn bitand(self, rhs: &'a UInt) -> Self::Output {
+    fn bitand(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_bitand(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> BitAnd<&'a UInt> for &'b mut UInt {
-    type Output = UInt;
+impl<'a, 'b> BitAnd<&'a Int> for &'b mut Int {
+    type Output = Int;
 
-    fn bitand(self, rhs: &'a UInt) -> Self::Output {
+    fn bitand(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_bitand(rhs).unwrap()
     }
 }
@@ -993,26 +1054,26 @@ impl<'a, 'b> BitAnd<&'a UInt> for &'b mut UInt {
 //  `BitOr` impls
 //  ===========================================================================
 
-impl<'a> BitOr<&'a UInt> for UInt {
-    type Output = UInt;
+impl<'a> BitOr<&'a Int> for Int {
+    type Output = Int;
 
-    fn bitor(self, rhs: &'a UInt) -> Self::Output {
+    fn bitor(self, rhs: &'a Int) -> Self::Output {
         self.into_bitor(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> BitOr<&'a UInt> for &'b UInt {
-    type Output = UInt;
+impl<'a, 'b> BitOr<&'a Int> for &'b Int {
+    type Output = Int;
 
-    fn bitor(self, rhs: &'a UInt) -> Self::Output {
+    fn bitor(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_bitor(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> BitOr<&'a UInt> for &'b mut UInt {
-    type Output = UInt;
+impl<'a, 'b> BitOr<&'a Int> for &'b mut Int {
+    type Output = Int;
 
-    fn bitor(self, rhs: &'a UInt) -> Self::Output {
+    fn bitor(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_bitor(rhs).unwrap()
     }
 }
@@ -1021,26 +1082,26 @@ impl<'a, 'b> BitOr<&'a UInt> for &'b mut UInt {
 //  `BitXor` impls
 //  ===========================================================================
 
-impl<'a> BitXor<&'a UInt> for UInt {
-    type Output = UInt;
+impl<'a> BitXor<&'a Int> for Int {
+    type Output = Int;
 
-    fn bitxor(self, rhs: &'a UInt) -> Self::Output {
+    fn bitxor(self, rhs: &'a Int) -> Self::Output {
         self.into_bitxor(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> BitXor<&'a UInt> for &'b UInt {
-    type Output = UInt;
+impl<'a, 'b> BitXor<&'a Int> for &'b Int {
+    type Output = Int;
 
-    fn bitxor(self, rhs: &'a UInt) -> Self::Output {
+    fn bitxor(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_bitxor(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> BitXor<&'a UInt> for &'b mut UInt {
-    type Output = UInt;
+impl<'a, 'b> BitXor<&'a Int> for &'b mut Int {
+    type Output = Int;
 
-    fn bitxor(self, rhs: &'a UInt) -> Self::Output {
+    fn bitxor(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_bitxor(rhs).unwrap()
     }
 }
@@ -1049,26 +1110,40 @@ impl<'a, 'b> BitXor<&'a UInt> for &'b mut UInt {
 //  `BitAndAssign`, `BitOrAssign` and `BitXorAssign` impls
 //  ===========================================================================
 
-impl<'a> BitAndAssign<&'a UInt> for UInt {
-    fn bitand_assign(&mut self, rhs: &'a UInt) {
+impl<'a> BitAndAssign<&'a Int> for Int {
+    fn bitand_assign(&mut self, rhs: &'a Int) {
         self.bitand_assign(rhs).unwrap();
     }
 }
 
-impl<'a> BitOrAssign<&'a UInt> for UInt {
-    fn bitor_assign(&mut self, rhs: &'a UInt) {
+impl<'a> BitOrAssign<&'a Int> for Int {
+    fn bitor_assign(&mut self, rhs: &'a Int) {
         self.bitor_assign(rhs).unwrap();
     }
 }
 
-impl<'a> BitXorAssign<&'a UInt> for UInt {
-    fn bitxor_assign(&mut self, rhs: &'a UInt) {
+impl<'a> BitXorAssign<&'a Int> for Int {
+    fn bitxor_assign(&mut self, rhs: &'a Int) {
         self.bitxor_assign(rhs).unwrap();
     }
 }
 
 /// # Arithmetic Operations
-impl UInt {
+impl Int {
+    /// Negates this `Int` inplace and returns the result.
+    /// 
+    /// **Note:** This will **not** allocate memory.
+    pub fn into_wrapping_neg(self) -> Int {
+        forward_mut_impl(self, Int::wrapping_neg)
+    }
+
+    /// Negates this `Int` inplace.
+    /// 
+    /// **Note:** This will **not** allocate memory.
+    pub fn wrapping_neg(&mut self) {
+        self.value.wrapping_neg()
+    }
+
     /// Adds `rhs` to `self` and returns the result.
     /// 
     /// **Note:** This will **not** allocate memory.
@@ -1076,8 +1151,8 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn into_wrapping_add(self, rhs: &UInt) -> Result<UInt> {
-        try_forward_bin_mut_impl(self, rhs, UInt::wrapping_add_assign)
+    pub fn into_wrapping_add(self, rhs: &Int) -> Result<Int> {
+        try_forward_bin_mut_impl(self, rhs, Int::wrapping_add_assign)
     }
 
     /// Add-assigns `rhs` to `self` inplace.
@@ -1087,7 +1162,7 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn wrapping_add_assign(&mut self, rhs: &UInt) -> Result<()> {
+    pub fn wrapping_add_assign(&mut self, rhs: &Int) -> Result<()> {
         self.value.wrapping_add_assign(&rhs.value)
     }
 
@@ -1101,8 +1176,8 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn into_wrapping_sub(self, rhs: &UInt) -> Result<UInt> {
-        try_forward_bin_mut_impl(self, rhs, UInt::wrapping_sub_assign)
+    pub fn into_wrapping_sub(self, rhs: &Int) -> Result<Int> {
+        try_forward_bin_mut_impl(self, rhs, Int::wrapping_sub_assign)
     }
 
     /// Subtract-assigns `rhs` from `self` inplace.
@@ -1115,7 +1190,7 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn wrapping_sub_assign(&mut self, rhs: &UInt) -> Result<()> {
+    pub fn wrapping_sub_assign(&mut self, rhs: &Int) -> Result<()> {
         self.value.wrapping_sub_assign(&rhs.value)
     }
 
@@ -1129,8 +1204,8 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn into_wrapping_mul(self, rhs: &UInt) -> Result<UInt> {
-        try_forward_bin_mut_impl(self, rhs, UInt::wrapping_mul_assign)
+    pub fn into_wrapping_mul(self, rhs: &Int) -> Result<Int> {
+        try_forward_bin_mut_impl(self, rhs, Int::wrapping_mul_assign)
     }
 
     /// Multiply-assigns `rhs` to `self` inplace.
@@ -1143,7 +1218,7 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn wrapping_mul_assign(&mut self, rhs: &UInt) -> Result<()> {
+    pub fn wrapping_mul_assign(&mut self, rhs: &Int) -> Result<()> {
         self.value.wrapping_mul_assign(&rhs.value)
     }
 
@@ -1158,8 +1233,8 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn into_wrapping_div(self, rhs: &UInt) -> Result<UInt> {
-        try_forward_bin_mut_impl(self, rhs, UInt::wrapping_div_assign)
+    pub fn into_wrapping_div(self, rhs: &Int) -> Result<Int> {
+        try_forward_bin_mut_impl(self, rhs, Int::wrapping_div_assign)
     }
 
     /// Assignes `self` to the division of `self` by `rhs`.
@@ -1173,8 +1248,8 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn wrapping_div_assign(&mut self, rhs: &UInt) -> Result<()> {
-        self.value.wrapping_udiv_assign(&rhs.value)
+    pub fn wrapping_div_assign(&mut self, rhs: &Int) -> Result<()> {
+        self.value.wrapping_sdiv_assign(&rhs.value)
     }
 
     /// Calculates the **unsigned** remainder of `self` by `rhs` and returns the result.
@@ -1188,8 +1263,8 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn into_wrapping_rem(self, rhs: &UInt) -> Result<UInt> {
-        try_forward_bin_mut_impl(self, rhs, UInt::wrapping_rem_assign)
+    pub fn into_wrapping_rem(self, rhs: &Int) -> Result<Int> {
+        try_forward_bin_mut_impl(self, rhs, Int::wrapping_rem_assign)
     }
 
     /// Assignes `self` to the **unsigned** remainder of `self` by `rhs`.
@@ -1203,8 +1278,37 @@ impl UInt {
     /// # Errors
     /// 
     /// - If `self` and `rhs` have unmatching bit widths.
-    pub fn wrapping_rem_assign(&mut self, rhs: &UInt) -> Result<()> {
-        self.value.wrapping_urem_assign(&rhs.value)
+    pub fn wrapping_rem_assign(&mut self, rhs: &Int) -> Result<()> {
+        self.value.wrapping_srem_assign(&rhs.value)
+    }
+}
+
+// ============================================================================
+//  Unary arithmetic negation: `std::ops::Add` and `std::ops::AddAssign`
+// ============================================================================
+
+impl Neg for Int {
+    type Output = Int;
+
+    fn neg(self) -> Self::Output {
+        self.into_wrapping_neg()
+    }
+}
+
+impl<'a> Neg for &'a Int {
+    type Output = Int;
+
+    fn neg(self) -> Self::Output {
+        self.clone().into_wrapping_neg()
+    }
+}
+
+impl<'a> Neg for &'a mut Int {
+    type Output = &'a mut Int;
+
+    fn neg(self) -> Self::Output {
+        self.wrapping_neg();
+        self
     }
 }
 
@@ -1212,24 +1316,24 @@ impl UInt {
 //  Add and Add-Assign: `std::ops::Add` and `std::ops::AddAssign`
 // ============================================================================
 
-impl<'a> Add<&'a UInt> for UInt {
-    type Output = UInt;
+impl<'a> Add<&'a Int> for Int {
+    type Output = Int;
 
-    fn add(self, rhs: &'a UInt) -> Self::Output {
+    fn add(self, rhs: &'a Int) -> Self::Output {
         self.into_wrapping_add(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> Add<&'a UInt> for &'b UInt {
-    type Output = UInt;
+impl<'a, 'b> Add<&'a Int> for &'b Int {
+    type Output = Int;
 
-    fn add(self, rhs: &'a UInt) -> Self::Output {
+    fn add(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_wrapping_add(rhs).unwrap()
     }
 }
 
-impl<'a> AddAssign<&'a UInt> for UInt {
-    fn add_assign(&mut self, rhs: &'a UInt) {
+impl<'a> AddAssign<&'a Int> for Int {
+    fn add_assign(&mut self, rhs: &'a Int) {
         self.wrapping_add_assign(rhs).unwrap()
     }
 }
@@ -1238,24 +1342,24 @@ impl<'a> AddAssign<&'a UInt> for UInt {
 //  Sub and Sub-Assign: `std::ops::Sub` and `std::ops::SubAssign`
 // ============================================================================
 
-impl<'a> Sub<&'a UInt> for UInt {
-    type Output = UInt;
+impl<'a> Sub<&'a Int> for Int {
+    type Output = Int;
 
-    fn sub(self, rhs: &'a UInt) -> Self::Output {
+    fn sub(self, rhs: &'a Int) -> Self::Output {
         self.into_wrapping_sub(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> Sub<&'a UInt> for &'b UInt {
-    type Output = UInt;
+impl<'a, 'b> Sub<&'a Int> for &'b Int {
+    type Output = Int;
 
-    fn sub(self, rhs: &'a UInt) -> Self::Output {
+    fn sub(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_wrapping_sub(rhs).unwrap()
     }
 }
 
-impl<'a> SubAssign<&'a UInt> for UInt {
-    fn sub_assign(&mut self, rhs: &'a UInt) {
+impl<'a> SubAssign<&'a Int> for Int {
+    fn sub_assign(&mut self, rhs: &'a Int) {
         self.wrapping_sub_assign(rhs).unwrap()
     }
 }
@@ -1264,24 +1368,24 @@ impl<'a> SubAssign<&'a UInt> for UInt {
 //  Mul and Mul-Assign: `std::ops::Mul` and `std::ops::MulAssign`
 // ============================================================================
 
-impl<'a> Mul<&'a UInt> for UInt {
-    type Output = UInt;
+impl<'a> Mul<&'a Int> for Int {
+    type Output = Int;
 
-    fn mul(self, rhs: &'a UInt) -> Self::Output {
+    fn mul(self, rhs: &'a Int) -> Self::Output {
         self.into_wrapping_mul(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> Mul<&'a UInt> for &'b UInt {
-    type Output = UInt;
+impl<'a, 'b> Mul<&'a Int> for &'b Int {
+    type Output = Int;
 
-    fn mul(self, rhs: &'a UInt) -> Self::Output {
+    fn mul(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_wrapping_mul(rhs).unwrap()
     }
 }
 
-impl<'a> MulAssign<&'a UInt> for UInt {
-    fn mul_assign(&mut self, rhs: &'a UInt) {
+impl<'a> MulAssign<&'a Int> for Int {
+    fn mul_assign(&mut self, rhs: &'a Int) {
         self.wrapping_mul_assign(rhs).unwrap();
     }
 }
@@ -1290,24 +1394,24 @@ impl<'a> MulAssign<&'a UInt> for UInt {
 //  Div and Div-Assign: `std::ops::Div` and `std::ops::DivAssign`
 // ============================================================================
 
-impl<'a> Div<&'a UInt> for UInt {
-    type Output = UInt;
+impl<'a> Div<&'a Int> for Int {
+    type Output = Int;
 
-    fn div(self, rhs: &'a UInt) -> Self::Output {
+    fn div(self, rhs: &'a Int) -> Self::Output {
         self.into_wrapping_div(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> Div<&'a UInt> for &'b UInt {
-    type Output = UInt;
+impl<'a, 'b> Div<&'a Int> for &'b Int {
+    type Output = Int;
 
-    fn div(self, rhs: &'a UInt) -> Self::Output {
+    fn div(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_wrapping_div(rhs).unwrap()
     }
 }
 
-impl<'a> DivAssign<&'a UInt> for UInt {
-    fn div_assign(&mut self, rhs: &'a UInt) {
+impl<'a> DivAssign<&'a Int> for Int {
+    fn div_assign(&mut self, rhs: &'a Int) {
         self.wrapping_div_assign(rhs).unwrap();
     }
 }
@@ -1316,24 +1420,24 @@ impl<'a> DivAssign<&'a UInt> for UInt {
 //  Rem and Rem-Assign: `std::ops::Rem` and `std::ops::RemAssign`
 // ============================================================================
 
-impl<'a> Rem<&'a UInt> for UInt {
-    type Output = UInt;
+impl<'a> Rem<&'a Int> for Int {
+    type Output = Int;
 
-    fn rem(self, rhs: &'a UInt) -> Self::Output {
+    fn rem(self, rhs: &'a Int) -> Self::Output {
         self.into_wrapping_rem(rhs).unwrap()
     }
 }
 
-impl<'a, 'b> Rem<&'a UInt> for &'b UInt {
-    type Output = UInt;
+impl<'a, 'b> Rem<&'a Int> for &'b Int {
+    type Output = Int;
 
-    fn rem(self, rhs: &'a UInt) -> Self::Output {
+    fn rem(self, rhs: &'a Int) -> Self::Output {
         self.clone().into_wrapping_rem(rhs).unwrap()
     }
 }
 
-impl<'a> RemAssign<&'a UInt> for UInt {
-    fn rem_assign(&mut self, rhs: &'a UInt) {
+impl<'a> RemAssign<&'a Int> for Int {
+    fn rem_assign(&mut self, rhs: &'a Int) {
         self.wrapping_rem_assign(rhs).unwrap();
     }
 }
@@ -1344,25 +1448,25 @@ impl<'a> RemAssign<&'a UInt> for UInt {
 
 use std::fmt;
 
-impl fmt::Binary for UInt {
+impl fmt::Binary for Int {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         self.value.fmt(f)
     }
 }
 
-impl fmt::Octal for UInt {
+impl fmt::Octal for Int {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         self.value.fmt(f)
     }
 }
 
-impl fmt::LowerHex for UInt {
+impl fmt::LowerHex for Int {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         self.value.fmt(f)
     }
 }
 
-impl fmt::UpperHex for UInt {
+impl fmt::UpperHex for Int {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         self.value.fmt(f)
     }
