+++ src/info/radix.rs
@@ -1,6 +1,5 @@
-
-use digit::{Digit, DigitRepr};
-use errors::{Error, Result};
+use crate::data::{Digit, DigitRepr};
+use crate::info::{Error, Result};
 
 /// A radix for parsing strings as `ApInt`s.
 /// 
@@ -112,7 +111,7 @@ impl Radix {
         // To generate this table:
         // ```
         //    for radix in 2u64..37 {
-        //        let mut power = digit::BITS / find_last_bit_set(radix.to_u8() as u64);
+        //        let mut power = Digit::BITS / find_last_bit_set(radix.to_u8() as u64);
         //        let mut base  = (radix.to_u8() as u32).pow(power as u32);
         //        while let Some(b) = base.checked_mul(radix) {
         //            base   = b;
