+++ src/logic.rs
@@ -1,0 +1,11 @@
+mod add;
+mod bitwise;
+mod div;
+mod fuzz;
+mod mul;
+mod cmp;
+mod shift;
+mod traits;
+mod utils;
+
+pub(crate) use self::utils::{forward_mut_impl, forward_bin_mut_impl, try_forward_bin_mut_impl};
