+++ src/logic/utils.rs
@@ -1,9 +1,9 @@
-use errors::Result;
+use crate::info::Result;
 
 /// Consumes `entity` and forwards it to an inplace-mutating function.
 /// 
 /// Returns the entity afterwards.
-pub fn forward_mut_impl<T, F>(entity: T, op: F) -> T
+pub(crate) fn forward_mut_impl<T, F>(entity: T, op: F) -> T
 where
     F: Fn(&mut T) -> ()
 {
@@ -15,7 +15,7 @@ where
 /// Consumes `entity` and forwards it to an inplace-mutating function.
 /// 
 /// Returns the entity afterwards.
-pub fn forward_bin_mut_impl<L, R, F>(entity: L, rhs: R, op: F) -> L
+pub(crate) fn forward_bin_mut_impl<L, R, F>(entity: L, rhs: R, op: F) -> L
 where
     F: Fn(&mut L, R) -> ()
 {
@@ -27,7 +27,7 @@ where
 /// Consumes `entity` and forwards it to an inplace-mutating function.
 /// 
 /// Returns the entity afterwards.
-pub fn try_forward_bin_mut_impl<L, R, F>(entity: L, rhs: R, op: F) -> Result<L>
+pub(crate) fn try_forward_bin_mut_impl<L, R, F>(entity: L, rhs: R, op: F) -> Result<L>
 where
     F: Fn(&mut L, R) -> Result<()>
 {
